Rename TsAst
The goal was to make lib.rs smaller and more readable.
